requirements.txt
#adafruit-blinka
#adafruit-circuitpython-dht
pillow
smbus2
luma.oled
luma.core
luma.lcd
RPi.GPIO
spidev
xmltodict
flask
waitress
pychromecast
gtts
zeep
pandas
pyproj
openpyxl
odfpy

# for LCD 20x4
#RPLCD
